r1.2, mkl=on, cuda=on, Sierra10.12.6

+++ tensorflow/core/kernels/variable_ops.h
@@ -180,10 +180,10 @@ class DestroyTemporaryVariableOp : public OpKernel {
     if (context->track_allocations()) {
       if (context->allocate_on_host(AllocatorAttributes())) {
         context->record_host_persistent_memory_allocation(
-            -static_cast<int64>(tmpvar.AllocatedBytes()));
+            -tmpvar.AllocatedBytes());
       } else {
         context->record_device_persistent_memory_allocation(
-            -static_cast<int64>(tmpvar.AllocatedBytes()));
+            -tmpvar.AllocatedBytes());
       }
     }
   }
